refactored envoyapi calls to use new set up in Slack listeners
