+++ apps/Envoy/routes/oauth-start.js
@@ -21,7 +21,7 @@ const startOAuthProcess = async (req, res) => {
 const _buildOAuthURL = () => {
     let auth_url = `${process.env.ENVOY_BASE_URL}/a/auth/v0/authorize`;
     auth_url += `?response_type=code&client_id=${process.env.ENVOY_CLIENT_ID}`;
-    auth_url += `&redirect_uri=${process.env.NGROK_URL}/oauthcallback&scope=locations.read+token.refresh`;
+    auth_url += `&redirect_uri=${process.env.NGROK_URL}/oauthcallback&scope=locations.read+token.refresh+flows.read+work-schedules.read+work-schedules.write+invites.read+invites.write+employees.read+reservations.read+reservations.write+sign-in-fields.read+sign-in-fields.write+sign-in-field-pages.read+badges.read`;
     return auth_url;
 };
